Add `is_supervised` to DEP profile
Although this field has been deprecated since iOS 13, omitting it in the Profile
causes an error at Apple's side. It is assumed that when omitted, it is set to
`false` and Apple's services return a `FLAGS_INVALID` error. For the sake of
compatibility, it is now included in the `Profile` class and set to `true` by default.